# docs/workflow/workflow-page-api-spec.yaml
openapi: 3.0.0
info:
  title: 工作流详情页 API
  version: 1.0.0
  description: |
    工作流详情页（/workflow/:id）相关的所有后端 API 接口定义。

    ## 页面功能概述

    工作流详情页是工作流编辑器的核心页面，支持以下功能：
    - 工作流基本信息管理（名称、描述、触发配置）
    - 工作流画布编辑（节点、连线、关联）
    - 工作流执行与状态监控
    - 执行历史查看
    - 工作流发布/取消发布

    ## API 分类

    1. **WorkflowController** - 工作流基础管理
    2. **WorkflowExecutionController** - 工作流执行管理
    3. **WorkflowNodeController** - 节点管理
    4. **WorkflowConnectionController** - 连线管理
    5. **WorkflowValidationController** - 验证与辅助接口
servers:
  - url: http://localhost:8080
    description: 后端服务地址
paths:
...
  /api/workflow/{id}:
    get:
      summary: 获取工作流详情
      description: |
        根据ID获取工作流的完整详细信息，包括：
        - 工作流基本信息（名称、描述、状态等）
        - 所有节点数据（nodes）
        - 所有连线数据（connections）
        - 所有关联数据（associations）

        ## 页面加载时调用
        访问 /workflow/:id 页面时，首先调用此接口加载工作流数据。
      operationId: getWorkflowDetail
      tags:
        - WorkflowController
      parameters:
        - name: id
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
            example: 30
      responses:
        '200':
          description: 工作流详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'

    put:
      summary: 更新工作流基本信息
      description: |
        更新工作流的基本信息，包括：
        - 名称（name）
        - 描述（description）
        - 触发类型（triggerType）
        - 触发配置（triggerConfig）

        ## 使用场景
        - 在工作流设置面板中修改名称/描述
        - 配置工作流触发方式（手动触发/定时触发/Webhook触发）
      operationId: updateWorkflow
      tags:
        - WorkflowController
      parameters:
        - name: id
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'

    delete:
      summary: 删除工作流
      description: |
        删除指定的工作流及其所有关联数据（节点、连线、执行记录等）。

        ## 使用场景
        - 在工作流列表或详情页点击删除按钮
      operationId: deleteWorkflow
      tags:
        - WorkflowController
      parameters:
        - name: id
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
...
  /api/workflows/{workflowId}/nodes/{nodeUuid}:
    get:
      summary: 获取特定节点
      description: |
        根据节点UUID获取节点详情。

        ## 使用场景
        - 单独刷新某个节点的数据
      operationId: getNode
      tags:
        - WorkflowNodeController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeUuid
          in: path
          required: true
          description: 节点UUID
          schema:
            type: string
      responses:
        '200':
          description: 节点详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeDTO'

    put:
      summary: 更新节点
      description: |
        更新指定节点的信息。

        ## 使用场景
        - 用户修改节点配置后保存
        - 用户移动节点位置后保存
      operationId: updateNode
      tags:
        - WorkflowNodeController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeUuid
          in: path
          required: true
          description: 节点UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeDTO'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeDTO'

    delete:
      summary: 删除节点
      description: |
        删除指定的节点。

        ## 注意事项
        - 删除节点时，相关的连线也会被删除
        - 如果是循环节点，关联的循环体画布也会被删除

        ## 使用场景
        - 用户选中节点后按Delete键或点击删除按钮
      operationId: deleteNode
      tags:
        - WorkflowNodeController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeUuid
          in: path
          required: true
          description: 节点UUID
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
...
  /api/workflows/{workflowId}/connections:
    get:
      summary: 获取工作流所有连线
      description: |
        获取指定工作流中的所有连线。

        ## 使用场景
        - 单独刷新连线数据
      operationId: getConnections
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 连线列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionDTO'

    post:
      summary: 创建连线
      description: |
        创建一条新的连线，连接两个节点的端口。

        ## 使用场景
        - 用户从源节点输出端口拖拽到目标节点输入端口
      operationId: createConnection
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionDTO'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionDTO'
  /api/workflows/{workflowId}/connections/{connectionUuid}:
    get:
      summary: 获取特定连线
      description: |
        根据连线UUID获取连线详情。
      operationId: getConnection
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: connectionUuid
          in: path
          required: true
          description: 连线UUID
          schema:
            type: string
      responses:
        '200':
          description: 连线详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionDTO'

    delete:
      summary: 删除连线
      description: |
        删除指定的连线。

        ## 使用场景
        - 用户选中连线后按Delete键或点击删除
      operationId: deleteConnection
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: connectionUuid
          in: path
          required: true
          description: 连线UUID
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
  /api/workflows/{workflowId}/connections/batch:
    post:
      summary: 批量创建连线
      description: |
        批量创建多条连线。

        ## 使用场景
        - 粘贴多个节点时同时创建相关连线
      operationId: batchCreateConnections
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ConnectionDTO'
      responses:
        '201':
          description: 批量创建成功

    delete:
      summary: 批量删除连线
      description: |
        批量删除多条连线。

        ## 使用场景
        - 删除多个节点时同时删除相关连线
      operationId: batchDeleteConnections
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                description: 连线UUID
      responses:
        '204':
          description: 批量删除成功
  /api/workflows/{workflowId}/connections/source/{nodeId}:
    get:
      summary: 获取从指定节点出发的连线
      description: |
        获取所有从指定节点（作为源节点）出发的连线。

        ## 使用场景
        - 分析节点的下游连接
      operationId: getConnectionsBySourceNode
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeId
          in: path
          required: true
          description: 源节点UUID
          schema:
            type: string
      responses:
        '200':
          description: 连线列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionDTO'
  /api/workflows/{workflowId}/connections/target/{nodeId}:
    get:
      summary: 获取到达指定节点的连线
      description: |
        获取所有到达指定节点（作为目标节点）的连线。

        ## 使用场景
        - 分析节点的上游连接
      operationId: getConnectionsByTargetNode
      tags:
        - WorkflowConnectionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeId
          in: path
          required: true
          description: 目标节点UUID
          schema:
            type: string
      responses:
        '200':
          description: 连线列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionDTO'

  # ==================== 验证与辅助接口 (WorkflowValidationController) ====================
...
components:
  schemas:
...
    WorkflowResponse:
      type: object
      description: 工作流完整响应
      properties:
        id:
          type: integer
          format: int64
          description: 工作流ID
          example: 30
        name:
          type: string
          description: 工作流名称
          example: "AI内容评测工作流"
        description:
          type: string
          description: 工作流描述
          example: "用于测试AI生成内容质量的自动化工作流"
        published:
          type: boolean
          description: 是否已发布
          example: true
        hasRun:
          type: boolean
          description: 是否已运行过
          example: true
        version:
          type: integer
          description: 版本号
          example: 3
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          description: 工作流状态
          example: "PUBLISHED"
        triggerType:
          type: string
          description: 触发类型
          enum: [MANUAL, SCHEDULE, WEBHOOK]
          example: "MANUAL"
        triggerConfig:
          type: string
          description: 触发配置（JSON格式）
          example: "{}"
        nodes:
          type: array
          description: 节点列表
          items:
            $ref: '#/components/schemas/NodeDTO'
        connections:
          type: array
          description: 连线列表
          items:
            $ref: '#/components/schemas/ConnectionDTO'
        associations:
          type: array
          description: 关联列表
          items:
            $ref: '#/components/schemas/AssociationDTO'
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
    WorkflowUpdateRequest:
      type: object
      description: 工作流更新请求
      properties:
        name:
          type: string
          description: 工作流名称
          example: "更新后的工作流名称"
        description:
          type: string
          description: 工作流描述
          example: "更新后的描述"
        triggerType:
          type: string
          enum: [MANUAL, SCHEDULE, WEBHOOK]
          description: 触发类型
          example: "MANUAL"
        triggerConfig:
          type: string
          description: 触发配置（JSON格式）
          example: "{}"
...
    NodeDTO:
      type: object
      description: |
        节点信息

        节点是工作流的基本组成单元，代表一个处理步骤。每个节点包含：
        - 唯一标识（nodeUuid）
        - 节点类型（type）
        - 位置信息（positionX, positionY）
        - 端口配置（inputPorts, outputPorts）
        - 参数定义（inputParams, outputParams）
        - 运行时配置（config）

        **注意**：`inputPorts`、`outputPorts`、`inputParams`、`outputParams`、`config` 字段都是 JSON 字符串格式。
      properties:
        id:
          type: integer
          format: int64
          description: 节点数据库ID（后端生成，保存时无需传入）
        nodeUuid:
          type: string
          description: |
            节点UUID，由前端生成的唯一标识符。
            格式示例：`node-start`、`node-loop`、`node-text-clean`
          example: "node-start"
        type:
          type: string
          description: |
            节点类型编码，常见类型：
            - `start`: 开始节点，定义工作流输入参数
            - `end`: 结束节点，定义工作流输出
            - `loop`: 循环节点，遍历数组执行循环体
            - `loopBodyCanvas`: 循环体画布，包含循环内部节点
            - `textClean`: 文本清洗节点
            - `tableExtract`: 表格提取节点
            - `tableGenerate`: 表格生成节点
            - `judgeModel`: 裁判模型节点
            - `apiAuto`: API调用节点
          example: "start"
        name:
          type: string
          description: 节点名称，用于显示和变量引用（如 `${开始.output}`）
          example: "开始"
        positionX:
          type: integer
          description: 节点在画布上的X坐标
          example: 80
        positionY:
          type: integer
          description: 节点在画布上的Y坐标
          example: 303
        inputPorts:
          type: string
          description: |
            输入端口配置，JSON字符串格式。
            示例：`[{"id":"input-1","name":"输入"}]`
          example: '[{"id":"input-1","name":"输入"}]'
        outputPorts:
          type: string
          description: |
            输出端口配置，JSON字符串格式。
            示例：`[{"id":"output-1","name":"输出"}]`
          example: '[{"id":"output-1","name":"输出"}]'
        inputParams:
          type: string
          description: |
            输入参数定义，JSON字符串格式。
            示例：`[{"name":"file","type":"File","fileType":"Excel","required":true}]`
          example: '[]'
        outputParams:
          type: string
          description: |
            输出参数定义，JSON字符串格式。
            示例：`[{"name":"output","type":"Array","elementType":"Object"}]`
          example: '[{"name":"dict","type":"Dictionary","required":false,"dictionaryType":"Demo数据字典"}]'
        config:
          type: string
          description: |
            节点运行时配置，JSON字符串格式。
            包含节点执行所需的具体参数值，支持变量引用。
          example: '{}'
        parentNodeUuid:
          type: string
          nullable: true
          description: |
            父节点UUID，用于嵌套节点（如循环体内部的节点指向循环体画布）。
            顶级节点此值为 null。
          example: null

    # ==================== 连线相关 ====================
    ConnectionDTO:
      type: object
      description: |
        连线信息

        连线定义节点之间的数据流向。从源节点的输出端口连接到目标节点的输入端口。
        连线的唯一标识（connectionUuid）由前端生成。
      properties:
        id:
          type: integer
          format: int64
          description: 连线数据库ID（后端生成，保存时无需传入）
        connectionUuid:
          type: string
          description: |
            连线UUID，由前端生成的唯一标识符。
            格式示例：`37ec2fbc-43bb-4a02-bcdc-28788790862e`
          example: "37ec2fbc-43bb-4a02-bcdc-28788790862e"
        sourceNodeUuid:
          type: string
          description: 源节点UUID（数据流出方）
          example: "node-start"
        sourcePortId:
          type: string
          description: 源节点的输出端口ID
          example: "output-1"
        targetNodeUuid:
          type: string
          description: 目标节点UUID（数据流入方）
          example: "node-text-clean"
        targetPortId:
          type: string
          description: 目标节点的输入端口ID
          example: "input-1"
        sourceParamIndex:
          type: integer
          nullable: true
          description: 源参数索引（用于多输出参数场景）
          example: null
        targetParamIndex:
          type: integer
          nullable: true
          description: 目标参数索引（用于多输入参数场景）
          example: null
        label:
          type: string
          nullable: true
          description: 连线标签（可选）
          example: null

    # ==================== 关联相关 ====================
    AssociationDTO:
      type: object
      description: |
        关联信息

        用于定义节点之间的特殊关联关系，目前主要用于：
        - 循环节点（loop）与循环体画布（loopBodyCanvas）的关联

        当删除循环节点时，需要同时删除关联的循环体画布。
      properties:
        id:
          type: integer
          format: int64
          description: 关联数据库ID（后端生成，保存时无需传入）
        loopNodeUuid:
          type: string
          description: 循环节点的UUID（关联的发起方）
          example: "node-loop"
        bodyNodeUuid:
          type: string
          description: 循环体画布节点的UUID（关联的目标方）
          example: "node-loop-body"
        associationType:
          type: string
          description: |
            关联类型，目前支持：
            - `loop-body`: 表示循环节点与循环体画布的关联
          example: "loop-body"

    # ==================== 执行相关 ====================
...